Auth forms refactor to use field components
